Refactor logic of EqualsIncompatibleType and extract it into a new file: TypeCompatibilityUtils.
RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=308121247
