+++ core/src/main/java/com/google/errorprone/bugpatterns/collectionincompatibletype/IncompatibleArgumentType.java
@@ -26,7 +26,8 @@ import com.google.errorprone.annotations.CheckReturnValue;
 import com.google.errorprone.annotations.CompatibleWith;
 import com.google.errorprone.bugpatterns.BugChecker;
 import com.google.errorprone.bugpatterns.BugChecker.MethodInvocationTreeMatcher;
-import com.google.errorprone.bugpatterns.EqualsIncompatibleType;
+import com.google.errorprone.bugpatterns.TypeCompatibilityUtils;
+import com.google.errorprone.bugpatterns.TypeCompatibilityUtils.TypeCompatibilityReport;
 import com.google.errorprone.matchers.Description;
 import com.google.errorprone.util.ASTHelpers;
 import com.google.errorprone.util.Signatures;
@@ -120,8 +121,8 @@ public class IncompatibleArgumentType extends BugChecker implements MethodInvoca
       Type argType = ASTHelpers.getType(argument);
       if (requiredType.type() != null) {
         // Report a violation for this type
-        EqualsIncompatibleType.TypeCompatibilityReport report =
-            EqualsIncompatibleType.compatibilityOfTypes(requiredType.type(), argType, state);
+        TypeCompatibilityReport report =
+            TypeCompatibilityUtils.compatibilityOfTypes(requiredType.type(), argType, state);
         if (!report.compatible()) {
           state.reportMatch(
               describeViolation(argument, argType, requiredType.type(), types, state));
